pyproject.toml
[project]
name = "bio2token"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "mamba-ssm>=2.2.2",
    "torch==2.4.1+cu121", # we have to specify this exactly, since we are pulling from specific indices and wheels later on
    "biopython>=1.84",
    "nglview>=3.1.2",
    "ipywidgets>=8.1.5",
    "jupyterlab>=4.2.5",
    "dill>=0.3.8",
    "invariant-point-attention>=0.2.2",
    "ipykernel>=6.29.5",
    "kaleido==0.2.1",
    "nbformat>=5.10.4",
    "pytest>=8.3.4",
    "seaborn>=0.13.2",
    "transformers",
    "mlflow>=2.19.0",
    "lightning>=2.5.0.post0",
    "hydra-zen>=0.14.0",
    "python-box>=7.3.2",
    "loguru>=0.7.3",
    "mmtf-python>=1.1.3",
    "biopython",
]

[tool.uv]
find-links = [
    "https://data.pyg.org/whl/torch-2.4.0%2Bcu121.html"
]
no-build-package = ["torch", "torchvision", "torchaudio", "torch-geometric", "torch-scatter", "torch-cluster", "mamba-ssm"]


[tool.uv.sources]
mamba_ssm = { url = "https://github.com/state-spaces/mamba/releases/download/v2.2.2/mamba_ssm-2.2.2+cu122torch2.4cxx11abiFALSE-cp311-cp311-linux_x86_64.whl" }
torch = { index = "pytorch" }
[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


